Cleaned BinaryToUnary Process and unit tests.
