+++ tests/lava/proc/event_data/utils.py
@@ -4,7 +4,6 @@
 
 import numpy as np
 import typing as ty
-import unittest
 
 from lava.magma.core.process.process import AbstractProcess
 from lava.magma.core.process.ports.ports import InPort, OutPort
@@ -15,12 +14,6 @@ from lava.magma.core.model.py.type import LavaPyType
 from lava.magma.core.resources import CPU
 from lava.magma.core.decorator import implements, requires
 from lava.magma.core.model.py.model import PyLoihiProcessModel
-from lava.magma.core.run_conditions import RunSteps
-from lava.magma.core.run_configs import Loihi1SimCfg
-from lava.proc.event_data.binary_to_unary_polarity.process \
-    import BinaryToUnaryPolarity
-from lava.proc.event_data.binary_to_unary_polarity.models \
-    import PyBinaryToUnaryPolarityPM
 
 
 class RecvSparse(AbstractProcess):
@@ -54,12 +47,14 @@ class PyRecvSparsePM(PyLoihiProcessModel):
     def run_spk(self) -> None:
         data, idx = self.in_port.recv()
 
-        self.data = np.pad(data,
-                           pad_width=(
-                               0, self.in_port.shape[0] - data.shape[0]))
-        self.idx = np.pad(idx,
-                          pad_width=(
-                              0, self.in_port.shape[0] - data.shape[0]))
+        self.data = np.pad(
+            data,
+            pad_width=(0, self.in_port.shape[0] - data.shape[0])
+        )
+        self.idx = np.pad(
+            idx,
+            pad_width=(0, self.in_port.shape[0] - data.shape[0])
+        )
 
 
 class SendSparse(AbstractProcess):
@@ -91,53 +86,4 @@ class PySendSparsePM(PyLoihiProcessModel):
         self._indices = proc_params["indices"]
 
     def run_spk(self) -> None:
-        data = self._data
-        idx = self._indices
-
-        self.out_port.send(data, idx)
-
-
-class TestProcessModelBinaryToUnaryPolarity(unittest.TestCase):
-    def test_init(self):
-        """Tests instantiation of the BinaryToUnary process model."""
-        pm = PyBinaryToUnaryPolarityPM()
-
-        self.assertIsInstance(pm, PyBinaryToUnaryPolarityPM)
-
-    def test_binary_to_unary_polarity_encoding(self):
-        """Tests whether the encoding from binary to unary works correctly."""
-        data = np.array([1, 1, 1, 0, 1, 0, 0, 1, 0])
-        indices = np.array([1, 5, 4, 3, 3, 2, 0, 1, 0])
-
-        expected_data = data
-        expected_data[expected_data == 0] = 1
-
-        expected_indices = indices
-
-        send_sparse = SendSparse(shape=(10, ), data=data, indices=indices)
-        binary_to_unary_encoder = BinaryToUnaryPolarity(shape=(10,))
-        recv_sparse = RecvSparse(shape=(10, ))
-
-        send_sparse.out_port.connect(binary_to_unary_encoder.in_port)
-        binary_to_unary_encoder.out_port.connect(recv_sparse.in_port)
-
-        run_cfg = Loihi1SimCfg()
-        run_cnd = RunSteps(num_steps=1)
-
-        send_sparse.run(condition=run_cnd, run_cfg=run_cfg)
-
-        sent_and_received_data = \
-            recv_sparse.data.get()[:expected_data.shape[0]]
-        sent_and_received_indices = \
-            recv_sparse.idx.get()[:expected_indices.shape[0]]
-
-        send_sparse.stop()
-
-        np.testing.assert_equal(sent_and_received_data,
-                                expected_data)
-        np.testing.assert_equal(sent_and_received_indices,
-                                expected_indices)
-
-
-if __name__ == '__main__':
-    unittest.main()
+        self.out_port.send(data=self._data, indices=self._indices)
